rename OxidationMaterials to OxidationPresets
